Added the current R MAMBI calculator package
Note that we still need to fix the way that the data is stored and accessed globally. We are close. We just need to make sure that we are using the correct function and then delete the portions that call the excel files.

Also, we need to add comments before each of the functions. To run the function, you use the following commands
library(devtools)
install() <- do this only if you have the package project open
library(MAMBIcalculator) <- note that the package name cannot have any underscores

Then you should be able to run the package as usual to debug.